# backend/api/openapi.yaml
swagger: "2.0"
host: localhost:8080
basePath: /
schemes:
  - http
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
security:
  - BearerAuth: []
definitions:
...
    Dish:
        properties:
            id:
                format: uint64
                type: integer
                x-go-name: ID
            diet_id:
                format: uint64
                type: integer
                x-go-name: DietID
            name:
                type: string
                x-go-name: Name
            calories:
                type: integer
                x-go-name: Calories
            protein:
                type: number
                format: float
                x-go-name: Protein
            carbs:
                type: number
                format: float
                x-go-name: Carbs
            fat:
                type: number
                format: float
                x-go-name: Fat
            created_at:
                format: date-time
                type: string
                x-go-name: CreatedAt
        type: object
        x-go-package: diet-app-backend/models
paths:
...
    /api/v1/diets/{dietId}/dishes:
        get:
            operationId: getDishes
            parameters:
                - description: Diet ID
                  format: uint64
                  in: path
                  name: dietId
                  required: true
                  type: integer
                  x-go-name: DietID
            responses:
                "200":
                    $ref: '#/responses/dishesResponse'
            summary: Returns all dishes for a diet.
            tags:
                - dishes
        post:
            operationId: createDish
            parameters:
                - description: Diet ID
                  format: uint64
                  in: path
                  name: dietId
                  required: true
                  type: integer
                  x-go-name: DietID
            responses:
                "201":
                    $ref: '#/responses/dishResponse'
            summary: Create a new dish for a diet.
            tags:
                - dishes
    /api/v1/diets/{dietId}/dishes/{id}:
        delete:
            operationId: deleteDish
            parameters:
                - description: Diet ID
                  format: uint64
                  in: path
                  name: dietId
                  required: true
                  type: integer
                  x-go-name: DietID
                - description: Dish ID
                  format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "204":
                    description: ""
            summary: Delete a dish.
            tags:
                - dishes
        put:
            operationId: updateDish
            parameters:
                - description: Diet ID
                  format: uint64
                  in: path
                  name: dietId
                  required: true
                  type: integer
                  x-go-name: DietID
                - description: Dish ID
                  format: uint64
                  in: path
                  name: id
                  required: true
                  type: integer
                  x-go-name: ID
            responses:
                "200":
                    $ref: '#/responses/dishResponse'
            summary: Update an existing dish.
            tags:
                - dishes
...
responses:
...
    dishResponse:
        description: ""
        schema:
            $ref: '#/definitions/Dish'
    dishesResponse:
        description: ""
        schema:
            items:
                $ref: '#/definitions/Dish'
            type: array
...